+++ extensions/vivaldi/src/components/NotInstalled.tsx
@@ -1,0 +1,48 @@
+import { useState } from "react";
+import { ActionPanel, Detail, showToast, Toast } from "@raycast/api";
+import { execSync } from "child_process";
+import { DEFAULT_ERROR_TITLE, DownloadVivaldiText } from "../constants";
+
+export function NotInstalled({
+  onInstall = () => {
+    return;
+  },
+}) {
+  const [isLoading, setIsLoading] = useState(false);
+  return (
+    <Detail
+      actions={
+        <ActionPanel>
+          {!isLoading && (
+            <ActionPanel.Item
+              title="Install with Homebrew"
+              onAction={async () => {
+                if (isLoading) return;
+
+                setIsLoading(true);
+
+                const toast = new Toast({ style: Toast.Style.Animated, title: "Installing..." });
+                await toast.show();
+
+                try {
+                  execSync(`brew install --cask vivaldi`);
+                  await toast.hide();
+                  onInstall();
+                } catch {
+                  await toast.hide();
+                  await showToast(
+                    Toast.Style.Failure,
+                    DEFAULT_ERROR_TITLE,
+                    "An unknown error occurred while trying to install"
+                  );
+                }
+                setIsLoading(false);
+              }}
+            />
+          )}
+        </ActionPanel>
+      }
+      markdown={DownloadVivaldiText}
+    />
+  );
+}