+++ extensions/vivaldi/CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add Vivaldi Open Tab action
 - Add Vivaldi Search Tabs action
 - Add Vivaldi Search History action
+- Add Vivaldi Search Bookmarks action